+++ src/all/background_page/model/entity/resource/resourceEntity.test.js
@@ -82,7 +82,45 @@ describe("Resource entity", () => {
       "folder_parent_id": "e2172205-139c-4e4b-a03a-933528123fff"
     });
     const contain = {secrets:true, permissions:true, permission:true, tags:true, favorite:true};
-    expect(entity.toDto(contain)).toEqual(dto);
+    expect(entity.toDto(contain)).toEqual({
+      "id": "10801423-4151-42a4-99d1-86e66145a08c",
+      "name": "test",
+      "username": "test@passbolt.com",
+      "uri": "https://www.passbolt.com",
+      "description": "Check check one two",
+      "deleted": false,
+      "created": "2020-05-04T20:31:45+00:00",
+      "modified": "2020-05-04T20:31:45+00:00",
+      "created_by": "d57c10f5-639d-5160-9c81-8a0c6c4ec856",
+      "modified_by": "d57c10f5-639d-5160-9c81-8a0c6c4ec856",
+      "favorite": {
+        "id": "45ce85c9-e301-4de2-8b41-298507002861",
+        "user_id": "d57c10f5-633d-5160-9c81-8a0c6c4ec856",
+        "foreign_key": "10801423-4151-42a4-99d1-86e66145a08c",
+        // "foreign_model": "Resource",
+        "created": "2020-05-06T21:59:24+00:00",
+        // "modified": "2020-05-06T21:59:24+00:00"
+      },
+      "secrets": [{
+        "id": "0dcde494-2231-43da-9bc5-6b39654b2a32",
+        "user_id": "d57c10f5-639d-5160-9c81-8a0c6c4ec856",
+        "resource_id": "10801423-4151-42a4-99d1-86e66145a08c",
+        "data": "-----BEGIN PGP MESSAGE-----\r\nVersion: OpenPGP.js v4.6.2\r\nComment: https:\/\/openpgpjs.org\r\n\r\nwcFMAxYTR81eetNbAQ\/\/TEWCA7W1kx7IzcZi4nmT92IZbdpzCBSQt5htSCoJ\r\nFfzGd27yeDT2GoEtmxmkG+gEak8ci0Jxa9FECaYDBzG4ixEDfDMfWqw\/WK2w\r\nj04oja+0qCAimV2nyItSYoaK5aZj8vL97V6U\/7YcraC9QTNY1Kd8RDPeL32D\r\nO2dpquPDLx5uMAmMoSZWruNCGqqJPjxMcxc2PBco+GJMcaGcYa5Y3+YueNpZ\r\nIIS0PbMpgiJlVvYzZywYC5lkIKFadVeV6MNkMmJfWB4VHq2Hoo3poZVP1rZV\r\n6cU7a7UuG4W3UUmezxQGQ6WAjh+qzkQHXrwI3cgU14du9sTCh8occwcPhG1C\r\nj8ljcTJqexQxA91TSj2UqhAnyB9yzZRcoh38bj\/OyGQmtiwxEFIzUymSi2pt\r\nysjJOZ7lB1Oh2l4vbgxJoNxtgvzY+3dsNXL510x793Hev3X2YcbO\/TJoy6G9\r\n89cuocJ1dlLIHqrfri43y1V0ZTfoa\/vigma4Qa5kUtB1tN0j38z+6tcjiz\/s\r\n8RJmXUK2bfHhvEbuc\/YnDDltpiZHc3QUtbj5TV2m+fO0ad2jVqxsi4eZid\/V\r\n\/WDUrAxRzY7xNRTRQQDbnT831NZeZbYobCpfPqU8ylF9iv\/V4lsyNYFrU0ne\r\n37JRFzl3cOY+jlqxGHaAF9\/mC3b3D3DmlZ+kOOQ7lE\/SwaoBAuDaJRsKzNqj\r\nTz8UFif5iwrEQY5BNzYd+zwGVzMlVP\/RNXR2YlAHx5lPMylgI73RDMoMZ4RT\r\nb7AQB9DqgobZI3dh3B90XqjkRiy3VJ\/nMhwknaZc6onJQgl2O\/ULie9kh69U\r\n1ojIkN+SHFCl42T1iT2eN08QUPffDVTMvT103WlX+MW8FV6CmF+TcDRUexs3\r\nT\/2EvFlxP6QTG41vLk4Sm3xce7rEZHiJ9hRrF26xVfT5jM+7z149lP5J8mgA\r\nARSBj2jlO7P1afQX+5RyYR+guD9LN95qMsNJwukTCzIo1AhE7yywf7b8v3a6\r\nXyanZo+TbDqxnJlozEMsdyGBwBn7UX6Erv072cZadO\/ZG2RBkbgiBGZ5hAjg\r\nPqwRAkfzDNa4WhsE9Crqs5ROy6IsDBGuAa8\/as0oCzIV+Ou4BPzKHfQDQS6U\r\nT0R+48sVAZAYY7TqaNHvf+3nlqMyssaK0SPm2fg3DZXPM2pcDatCFb4gVElC\r\n1qbG8pRIBmS\/NYr8m7IBnazDs9L6lYAjybuHes6cPqasDmHKha6DKl1P6jX+\r\nEeDxA0AVL4rZdUCt1fpEcFR\/R\/o4uDDLO8NGiHwM3MnbNI8G0SQy8q\/NhI11\r\nzWXyDeAR6hHKYC4h6WCCTFxe364PWLjQ5PGOLeAfeWEPCDZmP6U99kwoiOUu\r\ni8UuoIAFon3lIOXZnJ3ZtAcQ5UJ3gNcJH1EImZFdYtRgLo3GOPjBcNqGbmCu\r\n4xo+yMGy9Y8YJZM9HakKAChmHf01J3DAwNfUm8Rhx5w+NBQRm0aJ319wsACH\r\nlLEYvv+bVfPkNTvW\/vWND9eOPGI0Q8o=\r\n=AOt0\r\n-----END PGP MESSAGE-----\r\n",
+        "created": "2020-05-04T20:31:45+00:00",
+        "modified": "2020-05-04T20:31:45+00:00"
+      }],
+      "permission": {
+        "id": "d4c0e643-3967-443b-93b3-102d902c4510",
+        "aco": "Resource",
+        "aco_foreign_key": "10801423-4151-42a4-99d1-86e66145a08c",
+        "aro": "User",
+        "aro_foreign_key": "d57c10f5-639d-5160-9c81-8a0c6c4ec856",
+        "type": 15,
+        "created": "2020-05-04T20:31:45+00:00",
+        "modified": "2020-05-04T20:31:45+00:00"
+      },
+      "folder_parent_id": "e2172205-139c-4e4b-a03a-933528123fff"
+    });
   });
 
   it("constructor returns validation error if dto required fields are missing", () => {