+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [3.0.2] - 2021-01-27
+### Fixed
+- Fix do not enforce validation error for tags with slug duplicates
+- Fix do not enforce validation for avatar with empty user_id
 
-## [3.0.0] - 2021-01-27
+## [3.0.1] - 2021-01-27
 ### Fixed
 - Fix do not enforce validation error for tags with slug duplicates
 - Fix do not enforce validation for avatar with empty user_id